+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_CLIENT_ID=854591671315-j148rl3knd8t3j4tig9p3qhdpht4da91.apps.googleusercontent.com
