+++ client/components/studentList.js
@@ -1,0 +1,66 @@
+module.exports = [
+    {
+    firstName: 'Tony',
+    lastName: 'Robbins',
+    cohort: 'MCSP-07',
+    email: 'tonyrobbins@gmail.com',
+    gender: 'male',
+    terminalLeave: 'Aug 6',
+    etsDate: 'August 20, 2021',
+    etsCountDown: '20 days',
+    phoneNumber: '523-423-2551',
+    maritalStatus: 'Single',
+    deliverables: ['Complete capstone', 'CIF', 'Clear barracks'],
+    completedTasks: ['SFL TAP courses'],
+    currentPhase: 1,
+    selfieURL: 'https://i.pinimg.com/originals/35/5b/f7/355bf7c6116d655b605cbecf568223cc.jpg'
+  },
+  {
+    firstName: 'Johnny',
+    lastName: 'Cash',
+    cohort: 'MCSP-07',
+    email: 'cash@gmail.com',
+    gender: 'male',
+    terminalLeave: 'Aug 6',
+    etsDate: 'August 20, 2021',
+    etsCountDown: '20 days',
+    phoneNumber: '523-423-2551',
+    maritalStatus: 'Single',
+    deliverables: ['Complete capstone', 'CIF', 'Clear barracks'],
+    completedTasks: ['SFL TAP courses'],
+    currentPhase: 1,
+    selfieURL: 'https://i.pinimg.com/originals/35/5b/f7/355bf7c6116d655b605cbecf568223cc.jpg'
+  },
+  {
+    firstName: 'Shawn',
+    lastName: 'West',
+    cohort: 'MCSP-07',
+    email: 'cash@gmail.com',
+    gender: 'male',
+    terminalLeave: 'Aug 6',
+    etsDate: 'August 20, 2021',
+    etsCountDown: '20 days',
+    phoneNumber: '523-423-2551',
+    maritalStatus: 'Single',
+    deliverables: ['Complete capstone', 'CIF', 'Clear barracks'],
+    completedTasks: ['SFL TAP courses'],
+    currentPhase: 1,
+    selfieURL: 'https://i.pinimg.com/originals/35/5b/f7/355bf7c6116d655b605cbecf568223cc.jpg'
+  },
+  {
+    firstName: 'Suzie',
+    lastName: 'Q',
+    cohort: 'MCSP-07',
+    email: 'Suzie@hotmail.com',
+    gender: 'male',
+    terminalLeave: 'Aug 6',
+    etsDate: 'August 20, 2021',
+    etsCountDown: '20 days',
+    phoneNumber: '523-423-2551',
+    maritalStatus: 'Single',
+    deliverables: ['Complete capstone', 'CIF', 'Clear barracks'],
+    completedTasks: ['SFL TAP courses'],
+    currentPhase: 1,
+    selfieURL: 'https://i.pinimg.com/originals/35/5b/f7/355bf7c6116d655b605cbecf568223cc.jpg'
+  }
+];
